feature: add globalID and globalIDList methods to InputFieldBuilder
